+++ lib/widgets/task_tile.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+// リスト1つ分
+class TaskTile extends StatelessWidget {
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: Text('This is a task.'),
+      trailing: Checkbox(
+        value: false,
+        onChanged: (bool? value) {},
+      ),
+    );
+  }
+}
